+++ common/src/main/java/de/maxhenkel/voicechat/net/Channel.java
@@ -6,33 +6,37 @@ import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.server.network.ServerGamePacketListenerImpl;
 
-import java.util.ArrayList;
-import java.util.List;
+import javax.annotation.Nullable;
 
 public class Channel<T extends Packet<T>> {
 
-    private final List<NetManager.ClientReceiver<T>> clientListeners;
-    private final List<NetManager.ServerReceiver<T>> serverListeners;
+    @Nullable
+    private NetManager.ClientReceiver<T> clientListener;
+    @Nullable
+    private NetManager.ServerReceiver<T> serverListener;
 
     public Channel() {
-        clientListeners = new ArrayList<>();
-        serverListeners = new ArrayList<>();
+
     }
 
-    public void registerClientListener(NetManager.ClientReceiver<T> packetReceiver) {
-        clientListeners.add(packetReceiver);
+    public void setClientListener(NetManager.ClientReceiver<T> packetReceiver) {
+        clientListener = packetReceiver;
     }
 
-    public void registerServerListener(NetManager.ServerReceiver<T> packetReceiver) {
-        serverListeners.add(packetReceiver);
+    public void setServerListener(NetManager.ServerReceiver<T> packetReceiver) {
+        serverListener = packetReceiver;
     }
 
     public void onClientPacket(Minecraft client, ClientPacketListener handler, T packet) {
-        clientListeners.forEach(receiver -> receiver.onPacket(client, handler, packet));
+        if (clientListener != null) {
+            clientListener.onPacket(client, handler, packet);
+        }
     }
 
     public void onServerPacket(MinecraftServer server, ServerPlayer player, ServerGamePacketListenerImpl handler, T packet) {
-        serverListeners.forEach(receiver -> receiver.onPacket(server, player, handler, packet));
+        if (serverListener != null) {
+            serverListener.onPacket(server, player, handler, packet);
+        }
     }
 
 }