+++ forge/src/main/java/de/maxhenkel/voicechat/config/ForgeClientConfig.java
@@ -137,6 +137,10 @@ public class ForgeClientConfig extends ClientConfig {
                 .comment("If sound physics integration should be enabled")
                 .define("soundphysics", true)
         );
+        runLocalServer = wrapConfigEntry(builder
+                .comment("If voice chat should work in singleplayer/LAN worlds")
+                .define("run_local_server", true)
+        );
     }
 
 }